--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/assets/articulation/articulation.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/assets/articulation/articulation.py
@@ -13,8 +13,8 @@ from collections.abc import Sequence
 from prettytable import PrettyTable
 from typing import TYPE_CHECKING
 
-import carb
 import omni.isaac.core.utils.stage as stage_utils
+import omni.log
 import omni.physics.tensors.impl.api as physx
 from omni.isaac.core.utils.types import ArticulationActions
 from pxr import PhysxSchema, UsdPhysics
@@ -542,7 +542,7 @@ class Articulation(AssetBase):
         # update default joint pos to stay within the new limits
         if torch.any((self._data.default_joint_pos < limits[..., 0]) | (self._data.default_joint_pos > limits[..., 1])):
             self._data.default_joint_pos = torch.clamp(self._data.default_joint_pos, limits[..., 0], limits[..., 1])
-            carb.log_warn(
+            omni.log.warn(
                 "Some default joint positions are outside of the range of the new joint limits. Default joint positions"
                 " will be clamped to be within the new joint limits."
             )
@@ -925,13 +925,13 @@ class Articulation(AssetBase):
             raise RuntimeError(f"Failed to create articulation at: {self.cfg.prim_path}. Please check PhysX logs.")
 
         # log information about the articulation
-        carb.log_info(f"Articulation initialized at: {self.cfg.prim_path} with root '{root_prim_path_expr}'.")
-        carb.log_info(f"Is fixed root: {self.is_fixed_base}")
-        carb.log_info(f"Number of bodies: {self.num_bodies}")
-        carb.log_info(f"Body names: {self.body_names}")
-        carb.log_info(f"Number of joints: {self.num_joints}")
-        carb.log_info(f"Joint names: {self.joint_names}")
-        carb.log_info(f"Number of fixed tendons: {self.num_fixed_tendons}")
+        omni.log.info(f"Articulation initialized at: {self.cfg.prim_path} with root '{root_prim_path_expr}'.")
+        omni.log.info(f"Is fixed root: {self.is_fixed_base}")
+        omni.log.info(f"Number of bodies: {self.num_bodies}")
+        omni.log.info(f"Body names: {self.body_names}")
+        omni.log.info(f"Number of joints: {self.num_joints}")
+        omni.log.info(f"Joint names: {self.joint_names}")
+        omni.log.info(f"Number of fixed tendons: {self.num_fixed_tendons}")
 
         # container for data access
         self._data = ArticulationData(self.root_physx_view, self.device)
@@ -1142,7 +1142,7 @@ class Articulation(AssetBase):
                 velocity_limit=usd_velocity_limit[:, joint_ids],
             )
             # log information on actuator groups
-            carb.log_info(
+            omni.log.info(
                 f"Actuator collection: {actuator_name} with model '{actuator_cfg.class_type.__name__}' and"
                 f" joint names: {joint_names} [{joint_ids}]."
             )
@@ -1177,7 +1177,7 @@ class Articulation(AssetBase):
         # perform some sanity checks to ensure actuators are prepared correctly
         total_act_joints = sum(actuator.num_joints for actuator in self.actuators.values())
         if total_act_joints != (self.num_joints - self.num_fixed_tendons):
-            carb.log_warn(
+            omni.log.warn(
                 "Not all actuators are configured! Total number of actuated joints not equal to number of"
                 f" joints available: {total_act_joints} != {self.num_joints - self.num_fixed_tendons}."
             )
@@ -1336,7 +1336,7 @@ class Articulation(AssetBase):
                 effort_limits[index],
             ])
         # convert table to string
-        carb.log_info(f"Simulation parameters for joints in {self.cfg.prim_path}:\n" + table.get_string())
+        omni.log.info(f"Simulation parameters for joints in {self.cfg.prim_path}:\n" + table.get_string())
 
         # read out all tendon parameters from simulation
         if self.num_fixed_tendons > 0:
@@ -1372,4 +1372,4 @@ class Articulation(AssetBase):
                     ft_offsets[index],
                 ])
             # convert table to string
-            carb.log_info(f"Simulation parameters for tendons in {self.cfg.prim_path}:\n" + tendon_table.get_string())
+            omni.log.info(f"Simulation parameters for tendons in {self.cfg.prim_path}:\n" + tendon_table.get_string())