--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/assets/deformable_object/deformable_object.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/assets/deformable_object/deformable_object.py
@@ -9,7 +9,7 @@ import torch
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
-import carb
+import omni.log
 import omni.physics.tensors.impl.api as physx
 from pxr import PhysxSchema, UsdShade
 
@@ -305,7 +305,7 @@ class DeformableObject(AssetBase):
                         material_prim = mat_prim
                         break
         if material_prim is None:
-            carb.log_info(
+            omni.log.info(
                 f"Failed to find a deformable material binding for '{root_prim.GetPath().pathString}'."
                 " The material properties will be set to default values and are not modifiable at runtime."
                 " If you want to modify the material properties, please ensure that the material is bound"
@@ -341,14 +341,14 @@ class DeformableObject(AssetBase):
             self._material_physx_view = None
 
         # log information about the deformable body
-        carb.log_info(f"Deformable body initialized at: {root_prim_path_expr}")
-        carb.log_info(f"Number of instances: {self.num_instances}")
-        carb.log_info(f"Number of bodies: {self.num_bodies}")
+        omni.log.info(f"Deformable body initialized at: {root_prim_path_expr}")
+        omni.log.info(f"Number of instances: {self.num_instances}")
+        omni.log.info(f"Number of bodies: {self.num_bodies}")
         if self._material_physx_view is not None:
-            carb.log_info(f"Deformable material initialized at: {material_prim_path_expr}")
-            carb.log_info(f"Number of instances: {self._material_physx_view.count}")
+            omni.log.info(f"Deformable material initialized at: {material_prim_path_expr}")
+            omni.log.info(f"Number of instances: {self._material_physx_view.count}")
         else:
-            carb.log_info("No deformable material found. Material properties will be set to default values.")
+            omni.log.info("No deformable material found. Material properties will be set to default values.")
 
         # container for data access
         self._data = DeformableObjectData(self.root_physx_view, self.device)