--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/envs/direct_marl_env.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/envs/direct_marl_env.py
@@ -16,9 +16,9 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import Any, ClassVar
 
-import carb
 import omni.isaac.core.utils.torch as torch_utils
 import omni.kit.app
+import omni.log
 from omni.isaac.version import get_version
 
 from omni.isaac.lab.managers import EventManager
@@ -83,7 +83,7 @@ class DirectMARLEnv:
         if self.cfg.seed is not None:
             self.cfg.seed = self.seed(self.cfg.seed)
         else:
-            carb.log_warn("Seed not set for the environment. The environment creation may not be deterministic.")
+            omni.log.warn("Seed not set for the environment. The environment creation may not be deterministic.")
 
         # create a simulation context to control the simulator
         if SimulationContext.instance() is None:
@@ -105,7 +105,7 @@ class DirectMARLEnv:
                 f"({self.cfg.decimation}). Multiple multiple render calls will happen for each environment step."
                 "If this is not intended, set the render interval to be equal to the decimation."
             )
-            carb.log_warn(msg)
+            omni.log.warn(msg)
 
         # generate scene
         with Timer("[INFO]: Time taken for scene creation", "scene_creation"):