--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/envs/manager_based_env.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/envs/manager_based_env.py
@@ -8,8 +8,8 @@ import torch
 from collections.abc import Sequence
 from typing import Any
 
-import carb
 import omni.isaac.core.utils.torch as torch_utils
+import omni.log
 
 from omni.isaac.lab.managers import ActionManager, EventManager, ObservationManager
 from omni.isaac.lab.scene import InteractiveScene
@@ -78,7 +78,7 @@ class ManagerBasedEnv:
         if self.cfg.seed is not None:
             self.cfg.seed = self.seed(self.cfg.seed)
         else:
-            carb.log_warn("Seed not set for the environment. The environment creation may not be deterministic.")
+            omni.log.warn("Seed not set for the environment. The environment creation may not be deterministic.")
 
         # create a simulation context to control the simulator
         if SimulationContext.instance() is None:
@@ -106,7 +106,7 @@ class ManagerBasedEnv:
                 f"({self.cfg.decimation}). Multiple multiple render calls will happen for each environment step. "
                 "If this is not intended, set the render interval to be equal to the decimation."
             )
-            carb.log_warn(msg)
+            omni.log.warn(msg)
 
         # counter for simulation steps
         self._sim_step_counter = 0