--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/managers/manager_base.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/managers/manager_base.py
@@ -11,7 +11,7 @@ from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any
 
-import carb
+import omni.log
 
 import omni.isaac.lab.utils.string as string_utils
 from omni.isaac.lab.utils import string_to_callable
@@ -251,7 +251,7 @@ class ManagerBase(ABC):
                 if value.body_ids is not None:
                     msg += f"\n\tBody names: {value.body_names} [{value.body_ids}]"
                 # print the information
-                carb.log_info(msg)
+                omni.log.info(msg)
             # store the entity
             term_cfg.params[key] = value