--- a/source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/camera/camera.py
+++ b/source/extensions/omni.isaac.lab/omni/isaac/lab/sensors/camera/camera.py
@@ -251,7 +251,7 @@ class Camera(SensorBase):
             # TODO: Adjust to handle aperture offsets once supported by omniverse
             #   Internal ticket from rendering team: OM-42611
             if params["horizontal_aperture_offset"] > 1e-4 or params["vertical_aperture_offset"] > 1e-4:
-                carb.log_warn("Camera aperture offsets are not supported by Omniverse. These parameters are ignored.")
+                omni.log.warn("Camera aperture offsets are not supported by Omniverse. These parameters are ignored.")
 
             # change data for corresponding camera index
             sensor_prim = self._sensor_prims[i]